Add files via upload
This Excel VBA file contains a bunch of starter macros to introduce you to coding in VBA. The simplest (and arguably one of my most frequently-used) is TextToHyperlinks, which I showed in my May 2022 DFWTRN webinar: highlight a bunch of cells containing URLs that aren't hyperlinked (e.g., column output from scraping a people directory), double-click the macro to run it, and it prompts you to highlight the desired cells (if you hadn't done so already) and then instantly converts them to hyperlinks. You could also do this by creating a separate column with a =HYPERLINK() formula to reference the original column, but the macro saves you the time of doing that!